backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "minimarket-pos-backend"
version = "0.1.0"
description = "Backend para sistema POS e inventario de minimarket"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13,<2.0",
  "bcrypt>=4.1,<5.0",
  "email-validator>=2.1,<3.0",
  "fastapi>=0.110,<1.0",
  "passlib>=1.7,<2.0",
  "psycopg[binary]>=3.1,<4.0",
  "pydantic-settings>=2.2,<3.0",
  "python-jose[cryptography]>=3.3,<4.0",
  "python-multipart>=0.0.9,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "uvicorn[standard]>=0.29,<1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
